Button input delay fixed
Button input delay was a problem and was trying to figure out logic behind delay. Triggering login() on its own appears to have solved a problem I haven't been able to resolve by placing the text variable somewhere else within the function scope or making it global. May have to do with function structure, but still not sure. As it stands, it works without delay now.